Updates to enable changes made in the docker-splunk/multistage-build branch
Change default for build_location to be undefined. You can still set this value
to trigger an upgrade etc. Note that the old default value would never work with
existing playbooks because they expect that it has a specific filename format.

Fixed a bug with the version comparison logic when splunk.build_location is set
where it expected version numbers to always be single digit. This would break
for example for maintenance releases greater than '9', ie "7.2.10"

Added an update_etc.yml playbook which is used to run the /sbin/updateetc.sh
script included in the image. This is used to pour files from /opt/splunk-etc
over top /opt/splunk/etc, when an upgrade is necessary (if splunk.version differs).

Updated certain playbooks to only be executed on first time run (ftr)

Fixed a few cases where "splunk.<something> is defined" never works as expected