By default, the ansible "uri" module does not urlencode body content. There were many places in playbooks where it was being used with variables plugged directly into the body content, which led to the possibility of it sending invalid HTTP requests to the Splunk REST APIs.
For example, randomly generated secrets that contained a % character resulted
in failures like this one:

Status code was 400 and not [201, 409]: HTTP Error 400: Unparsable URI-encoded request data

This could easily be reproduced by including invalid characters in shc.secret.

To fix this, I updated almost all of the instances of "uri" to use "form-urlencoded"
for "body_format" and splitting out each key/value pair inside the body object.
This way, ansible is always used to properly urlencoded the requests being sent
to the Splunk Rest APIs.