Moved splunk_search_head_captain fact from the top-level site.yml to the main.yml of the splunk_search_head role. It was out of place in site.yml and is only used by plays within that role.
Added "Add new member to SHC" which can be used to append new
search heads to an existing cluster. Previously, it was not possible
to add new members.

Sanity check to allow destructive sync to fail if member is the
captain. This was previously causing the playbooks to repeatedly
retry the command until failing, whenever the captaincy
was transferred to a member other than the original one, and then
re-run on that member (i.e. due to restart).

Reduced destructive sync retry delay from 60 seconds (which seemed
quite excessive) to 10 seconds.