Updated wait_for_splunk_instance.yml so that the HTTP request it makes does not require authentication. When waiting for a remote instance, this created a dependency that the remote instance must have the same admin password as the local instance, which likely doesn't hold true outside of crafted test environments. Splunk itself has no such requirement, which is why it has and uses pass4SymmKey secrets.
Only the root endpoint appears to be unauthenticated in the REST API,
so this prevents us from being able to query uptime of the remote
instance and ensure it is > 10 seconds. I don't know why this check was
put in place, but generally artificial wait times like this can be
better handled via more specific checks and retry loops.