Added license_master_url and cluster_master_url parameters to the splunk section of the default.yml config. These may be used to define the location of these components via YAML, without having to use the LICENSE_MASTER_URL and CLUSTER_MASTER_URL environment variables
Note that the environment variables still work the same as before.
This just makes it easier to configure them. Previously, to use a
license master, one had to specify the location via an environment
variable and the secret (pass4SymmKey) via default.yml. This just
makes it a lot easier for users. For example, one can now create
a "license_master.yml" config that only includes these parameters
and layer it in with other configs:

splunk:
  license_master_url: my-license-master
  pass4SymmKey: P@ssw0rd

Or, create a "cluster_master.yml" config:

splunk:
  cluster_master_url: my-license-master
  idxc:
    pass4SymmKey: P@ssw0rd

This PR also includes a change for the standalone instance role
so that it works as a standalone search head when a cluster master
is defined. Previously, any cluster master config was ignored when
running in the standalone role.