Add apps_location_local for local apps on CM/Deployer (#628)
* Add apps_location_local for local apps on CM/Deployer

For the CM and deployer, apps specified in `apps_location` are put
into the bundle and pushed to the cluster and are disabled on
the local CM/Deployer instance.  This commit add an
`apps_location_local` parameter, much like `apps_location` that
can be used to specify apps to be installed locally on the CM or
deployer.

Specifying this parameter in the default.yaml would be the same
as `apps_location` and both could be used concurrently in a
given configuration.  For example on a CM the following config
would install apps 1 and 2 clusterwide and app 3 locally only:

```
  splunk:
    apps_location:
      - "/mnt/apps/application1.tgz"
      - "/mnt/apps/application2.tgz"
    apps_location_local:
      - "/mnt/apps/application3.tgz"
```