Add app_paths_install variable for app install
Add app_paths_install variable to install apps directly into the app_paths location
for the default (local), idxc, shc, and deployment roles.  The syntax is:

      app_paths_install:
        shc:
          - "/mnt/apps/application1.tgz"
        idxc:
          - "/mnt/apps/application2.tgz"
        default:
          - "/mnt/apps/application3.tgz"

Apps listed under default will be installed locally to the location specified in:
  app_paths:
    default:
which typically is etc/apps.  Similarly for idxc (etc/master-apps) on CM,
shc (etc/shcluster/apps) on the deployer, and deployment (etc/deployment-apps) on
the deployment server.  While this is meant as a replacement for apps_location,
the funtionality of apps_location is untouched and either could be used.