Add app_paths_install variable for app install (#632)
* Add app_paths_install variable for app install

Add app_paths_install variable to install apps directly into the app_paths location
for the default (local), idxc, shc, and deployment roles.  The syntax is:

      app_paths_install:
        shc:
          - "/mnt/apps/application1.tgz"
        idxc:
          - "/mnt/apps/application2.tgz"
        default:
          - "/mnt/apps/application3.tgz"

Apps listed under default will be installed locally to the location specified in:
  app_paths:
    default:
which typically is etc/apps.  Similarly for idxc (etc/master-apps) on CM,
shc (etc/shcluster/apps) on the deployer, and deployment (etc/deployment-apps) on
the deployment server.  While this is meant as a replacement for apps_location,
the funtionality of apps_location is untouched and either could be used.

* Address PR comments

 - Add documentation for `app_paths_install` config options
 - Add unit tests
 - Add suport for URL and splunkbase retrieval

* Update default.yml.md docs

* Address PR comments for backwards compatibility

* Add check for uninitialized variable

* Set apps_location_local to NULL.

* Add check for app_list existence

* Revert unnecessary changes

* Change base debian10 image for molecule

The tests require that the molecule image for debian10 contains the
packages: procps, curl, and net-tools.  Change to a base image that
contains these packages.