CSPL-1368 Ensure splunk-ansible gracefully handles ES app upgrades
Currently when attempting to install a new version of ES over an older version of ES through ansible on a SHC deployer or Cluster Manager, the "Run ESS post-install setup" task which is responsible for running the essinstall command will fail.

The failure occurs because ansible will disable the SplunkEnterpriseSecurity app on the deployer and cluster manager instances as one of the last steps of the initial install. This means that when the "Run ESS post-install setup" is attempted during the upgrade, the SplunkEnterpriseSecuritySuite app is in the disabled state, and the essinstall command cannot be found and therefore executed.

Fix: The simple fix for this is to add one more task to configure_ess.yml to ensure the SplunkEnterpriseSecuritySuite app is enabled prior to attempting to run essinstall.