CSPL-2088: [Splunk 9.0.x]KV store fails to come up on a pod reset (#700)
When a Pod resets, a new pod comes up and uses the same persistent
`etc` and `var` locations. Pod reset causes the file permissions to
change. Open permissions on `var/lib/splunk/kvstore/mongo/splunk.key`
causing the Mongod 4.2.17 to come up during the migration path, that
way the mongod is marked to start with 3.6.x, and that fails forever
as the existing mongod db is already 4.2.x compatiable.